AQ-22322 Add ExampleFieldDataPlugins
Created a solution containing two example field data plugins to demonstrate how to write a plugin for the field data plugin framework.  The plugins compile, but they do not import data because it is missing functionality to read and parse a field data file.  This functionality is stubbed out in FileReader.cs.