PF-70 Improved the Tableau WebConnector
1) Restore the selected time-series when refreshing the data

When a connection is edited, restore the previously selected time-series and time-range.

This improves the usability when a saved Tableau workbook is reopened two weeks later.

2) Added the "To Time" input value

Now a specific time-range of points can be retrieved.

3) Added the "Time Align Points?" checkbox

This boolean value defaults to true, and will perform time-alignment to the first time-series in the selected list.
When false, each time-series will be retrieved at its native resolution

4) Added support for HTTPS connections

If no scheme is specified, "http://" is assumed

5) Updated the readme