PF-179 Use short-lived SOS client connections to ensure that changes are committed
We can't use a single SOS session to perform all the work.
The transactions are journaled and need to be committed more frequently.

Some test runs had a 30-hour export for the initial sync, and would eventually cause a DeleteSensor operation to take 8 hours to complete.
Using short-lived SOS sessions solves this.