Issue-94 Added more command line options for time-series creation
Added ability to set unit, comment, computation and sublocation identifiers, publish flag, and extended attributes when creating a time-series.

When the -SourceTimeSeries option is used, use the source time-series properties as much as possible if a new time-series is being created.

This brings PointZilla closer to cloning the corrected value of a time-series. Some settings are still not a perfect match, so it isn't recommend for officially migrating a time-series. But it is pretty close.

PointZilla is now a 64-bit-only app so it can deal with bigger signals on modern systems.