Fix response_or_raise handler to properly report API response message.
response.json is a function and requires the brackets to invoke (otherwise you are passing the function), and simplify the check for attribute existence. This results bug in not correctly extracting the error message response from the API, so calls that fail do not receive proper details on why they fail - just an HTTP error code and standard text.