PF-1417 - Added more support for NWIS text files
1) Added -CsvSkipRowsAfterHeader= option (to skip the dummy line after the header in NWIS data)

2) Added regex field matching support

The NWIS data columns have a name like "1234_00060", where 1234 is the timeseries ID and 00060 is the parameter code.

Allow field names to be surround by forward slashes to indicate a regular express match.

"/_00060/" will match the first column that ends with the 00060 parameter code for discharge.

"/_00060/#2" will match the 2nd discharge parameter column.

3) Added the -CsvFormat="NWIS" option

This will also populate USGS timezone aliases.